a lot of adjustments

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: feandrad <feandrad@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/08 00:34:07 by feandrad          #+#    #+#             */
-/*   Updated: 2022/09/25 23:50:11 by feandrad         ###   ########.fr       */
+/*   Updated: 2022/09/30 03:17:56 by feandrad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int	result;
+	size_t	result;
 
 	result = 0;
 	while (str[result])
